2025/1/4

+++ Lampyris OKX Trading Server/Sources/Utilities/DateTimeUtil.cs
@@ -37,33 +37,33 @@ public static class DateTimeUtil
         switch (okxBarSize)
         {
             case OkxBarSize._1m:
-                return timeSpan.TotalMinutes;
+                return (int)(timeSpan.TotalMinutes);
             case OkxBarSize._3m:
-                return timeSpan.TotalMinutes / 3;
+                return (int)(timeSpan.TotalMinutes / 3);
             case OkxBarSize._5m:
-                return timeSpan.TotalMinutes / 5;
+                return (int)(timeSpan.TotalMinutes / 5);
             case OkxBarSize._15m:
-                return timeSpan.TotalMinutes / 15;
+                return (int)(timeSpan.TotalMinutes / 15);
             case OkxBarSize._30m:
-                return timeSpan.TotalMinutes / 30;
+                return (int)(timeSpan.TotalMinutes / 30);
             case OkxBarSize._1H:
-                return timeSpan.TotalHours;
+                return (int)(timeSpan.TotalHours);
             case OkxBarSize._2H:
-                return timeSpan.TotalHours / 2;
+                return (int)(timeSpan.TotalHours / 2);
             case OkxBarSize._4H:
-                return timeSpan.TotalHours / 4;
+                return (int)(timeSpan.TotalHours / 4);
             case OkxBarSize._6H:
-                return timeSpan.TotalHours / 6;
+                return (int)(timeSpan.TotalHours / 6);
             case OkxBarSize._12H:
-                return timeSpan.TotalHours / 12;
+                return (int)(timeSpan.TotalHours / 12);
             case OkxBarSize._1D:
-                return timeSpan.TotalDays;
+                return (int)(timeSpan.TotalDays);
             case OkxBarSize._2D:
-                return timeSpan.TotalDays / 2;
+                return (int)(timeSpan.TotalDays / 2);
             case OkxBarSize._3D:
-                return timeSpan.TotalDays / 3;
+                return (int)(timeSpan.TotalDays / 3);
             case OkxBarSize._1W:
-                return timeSpan.TotalDays / 7;
+                return (int)(timeSpan.TotalDays / 7);
             case OkxBarSize._1M:
                 return Math.Abs(lhs.Year * lhs.Month - rhs.Year * rhs.Month);
             case OkxBarSize._3M:
